ai-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
pytesseract
Pillow
spacy
nltk
python-multipart
